test: move to proper integration tests
Correctly build the git-perf bin as a dependency and use the correct
version of the binary in the integration tests

topic: move-to-proper-integration-tests
